Day 2

+++ 2020/01/part2.jl
@@ -1,5 +1,7 @@
+#https://adventofcode.com/2020/day/1
+
 open("input.txt") do file
-    expenses = [parse(Int64, line) for line in eachline(file)]
+    expenses = [parse(Int, line) for line in eachline(file)]
     for i in 1:size(expenses, 1)
         for j in i+1:size(expenses, 1)
             for k in j+1:size(expenses, 1)
